handling missing jenkins archive

+++ VMs-master-worker-target/scripts/docker-jenkins.sh
@@ -9,12 +9,29 @@ if [ -d "$pathToJenkinsHome/jenkins_home" ]
 then 
     echo " Folder jenkins_home exist and skip the preparation."
 else
-    echo "Preparing jenkins_home for swapping"
-    cd $pathToJenkinsHome
-    cat  jenkins-home-tar-gz* > jenkins_home.tar.gz
-    tar -xvzf jenkins_home.tar.gz 
-    rm jenkins_home.tar.gz
-    echo "Folder jenkins_home is prepared"
+    if [ -f jenkins-home-tar-gzaa ]
+    then
+        echo "Preparing jenkins_home for swapping"
+        cd $pathToJenkinsHome
+        cat  jenkins-home-tar-gz* > jenkins_home.tar.gz
+        tar -xvzf jenkins_home.tar.gz 
+        rm jenkins_home.tar.gz
+        rm jenkins-home-tar*
+        rm split-backup.sh
+        echo "Folder jenkins_home is prepared"
+    else
+        echo "****************************"
+        echo
+        echo "Jenkins Archive file is missing or broken. "
+        echo "Folder jenkins_home will be not configure for you."
+        echo "Please, Reclone project or configure manually"
+        echo
+        echo " Your initial password string: "
+        docker exec myjenkins bash -c "cat  /var/jenkins_home/secrets/initialAdminPassword"
+        echo
+        echo "********************************"
+    fi    
+
 fi
 echo "Swapping jenkins_home for running myjenkins container"
 sleep 10
